Add neo4j-migrations (#19)
This change adds neo4j-migrations to the labs project.
The project fullfils all requirements for the labs project and is used
in several places internally and externally.

I devote a fixed amount of time on a regular basis to improve and
maintain it. The 1.x version has been released last months and we
guarantee a stable API for it now.

It uses GH actions for releases, has over 80% check coverage and a
quadruple A rating on Sonarcloud with standard settings.

The docs links point to the readme, a full blown documentation is
planned for the end of this year.